pyproject.toml
[tool.poetry]
name = "telegram-work"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "telegram_work"}]

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^2.25.1"
django-environ = "^0.10.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
